first dat stage 2

+++ src/main/java/com/couponmania2/coupon_project/repositories/CompanyRepo.java
@@ -3,5 +3,5 @@ package com.couponmania2.coupon_project.repositories;
 import com.couponmania2.coupon_project.beans.Company;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CompanyRepo extends JpaRepository <Company , Long> {
+public interface CompanyRepo extends JpaRepository <Company , Integer> {
 }
